Выделение функции main() в отдельный модуль рантайма (#170)
