+++ autotests/Hash/run.bat
@@ -29,6 +29,7 @@ setlocal
     refalrts-debugger ^
     refalrts-dynamic ^
     refalrts-functions ^
+    refalrts-main ^
     refalrts-profiler ^
     refalrts-vm ^
     refalrts-platform-specific
