+++ autotests/Library/run.bat
@@ -30,6 +30,7 @@ setlocal
     refalrts-debugger ^
     refalrts-dynamic ^
     refalrts-functions ^
+    refalrts-main ^
     refalrts-profiler ^
     refalrts-vm ^
     refalrts-platform-specific
