+++ autotests/compatibility/run.bat
@@ -97,6 +97,7 @@ goto :EOF
     refalrts-debugger ^
     refalrts-dynamic ^
     refalrts-functions ^
+    refalrts-main ^
     refalrts-profiler ^
     refalrts-vm ^
     refalrts-platform-specific 2>__error.txt
