+++ autotests/compatibility/run.sh
@@ -106,6 +106,7 @@ prepare_prefix() {
     refalrts-debugger \
     refalrts-dynamic \
     refalrts-functions \
+    refalrts-main \
     refalrts-profiler \
     refalrts-vm \
     refalrts-platform-POSIX \
