+++ autotests/lexgen/run.bat
@@ -52,6 +52,7 @@ setlocal
     refalrts-debugger ^
     refalrts-dynamic ^
     refalrts-functions ^
+    refalrts-main ^
     refalrts-profiler ^
     refalrts-vm ^
     refalrts-platform-specific
