+++ autotests/lexgen/run.sh
@@ -99,6 +99,7 @@ run_test() {
     refalrts-debugger
     refalrts-dynamic
     refalrts-functions
+    refalrts-main
     refalrts-profiler
     refalrts-vm
     refalrts-platform-POSIX
