+++ autotests/run.bat
@@ -81,6 +81,7 @@ setlocal
     refalrts-debugger ^
     refalrts-dynamic ^
     refalrts-functions ^
+    refalrts-main ^
     refalrts-profiler ^
     refalrts-vm ^
     refalrts-platform-specific
