+++ autotests/run.sh
@@ -174,6 +174,7 @@ run_test() {
     refalrts-debugger
     refalrts-dynamic
     refalrts-functions
+    refalrts-main
     refalrts-profiler
     refalrts-vm
     refalrts-platform-POSIX
