+++ src/srlib-rich-prefix/rich-prefix.sref
@@ -7,6 +7,7 @@
 //FROM refalrts-debugger
 //FROM refalrts-dynamic
 //FROM refalrts-functions
+//FROM refalrts-main
 //FROM refalrts-platform-specific
 //FROM refalrts-profiler
 //FROM refalrts-vm
