+++ src/srlib-slim-prefix/slim-prefix.sref
@@ -5,6 +5,7 @@
 //FROM refalrts-debugger
 //FROM refalrts-dynamic
 //FROM refalrts-functions
+//FROM refalrts-main
 //FROM refalrts-platform-specific
 //FROM refalrts-profiler
 //FROM refalrts-vm
