+++ src/srlib/make.sh
@@ -76,6 +76,7 @@ prepare_common() {
     refalrts-debugger
     refalrts-dynamic
     refalrts-functions
+    refalrts-main
     refalrts-platform-specific
     refalrts-profiler
     refalrts-vm
